desc.realizer ready
