+++ python/__init__.py
@@ -1,2 +1,2 @@
 import pkgutil
-__path__ = pkgutil.extend_path(__path__,__name__)
+__path__ = pkgutil.extend_path(__path__, __name__)
