+++ python/desc/slrealizer/SLRealizer.py
@@ -3,77 +3,108 @@ import pandas as pd
 import matplotlib.pyplot as plt
 import pylab
 import matplotlib
+import math
 
 class SLRealizer(object):
-    '''
-    Example class for SLRealizer package.
-    '''
+    
+    scale_factor = 2
+    
+    # Maybe make a separate method for it
+    # Best OOP practice?
+
     def __init__(self, lens = None, observation = "../../data/twinkles_observation_history.csv"):
         """
             Reads in a lens sample catalog and observation data.
             We assume lenses are OM10 lenses and observation file is .csv file
         """
         self.lens = lens
-        self.observation = observation
-        self.currObs = None
-        self.currLens = None
-        obsHist = pd.read_csv(observation,index_col=0).as_matrix()
+        self.observation = pd.read_csv(observation,index_col=0).as_matrix()
     
-    def plot_random_lens_date(lensID = 7176527):
-        import random
-        randomIndex = random.randint(0, 200)
-        if(self.observation[randomIndex][1]!='y'):
-            self.currObs = self.observation[randomIndex]
-            self.currLens = self.lens.get_lens(lensID)
-            #obsHist has MJD Filter FWHM 5sigmag
-            filterQuasar = self.currObs[1] + '_SDSS_quasar'
-            filterLens = self.currObs[1] + '_SDSS_lens'
-            lens_mag = self.currLens[filterLens]
-            quasar_mag = self.currLens[filterQuasar]
-            mag_ratio = math.pow(2.5, -lens_mag+quasar_mag)
-            if(mag_ratio>1):
-                quasar_alpha = 1/mag_ratio
-                lens_alpha = 1
-            else:
-                quasar_alpha = 1
-                lens_alpha = mag_ratio
-            scale_factor = 2
-            sourceX = self.lens['XSRC'][0]
-            sourceY = self.lens['YSRC'][0]
-            lensX = self.lens['XIMG'][0]
-            lensY = self.lens['YIMG'][0]
-            if (abs(lensX[0]+sourceX))>(abs(sourceX)):
-                plotX = abs(lensX[0]+sourceX)
-            else:
-                plotX = abs(sourceX)
-            if (abs(lensY[0]+sourceY))>(abs(sourceY)):
-                plotY = abs(lensY[0]+sourceY)
-            else:
-                plotY = abs(sourceY)
-                fig = plt.figure(figsize=(4, 16))
-            for i in range(4):
-                if(obsHist[1]=='g'):
-                    circleColor = 'g'
-                else:
-                    circleColor = 'r'
-                plotNumStr = "4"+"1"+str(i+1)
-                plotNum = int(plotNumStr)
-                sub = fig.add_subplot(plotNum)
-                sub.set_ylim(-plotY-2*obsHist[2]/scale_factor, plotY+2*obsHist[2]/scale_factor)
-                sub.set_xlim(-plotX-2*obsHist[2]/scale_factor, plotX+2*obsHist[2]/scale_factor)
-                source = plt.Circle((sourceX, sourceY), radius=obsHist[2]/scale_factor, alpha=quasar_alpha, fc=circleColor, linewidth=0)
-                lens = plt.Circle((lensX[i]+sourceX, lensY[i]+sourceY), radius=obsHist[2]/scale_factor, alpha=lens_alpha, fc=circleColor, linewidth=0)
-                fig.gca().add_patch(lens)
-                fig.gca().add_patch(source)
-                sub.set_title('Observation ' + str(i+1) + ' with ' + 'filter ' + obsHist[1])
-                seeing = plt.Circle(((-plotY-2*obsHist[2])/scale_factor, (plotX-2*obsHist[2])/scale_factor), radius=obsHist[2]/scale_factor, alpha=0.1, fc='k')
-                sub.legend((source, seeing, lens),('source', 'seeing', 'lens'))
-                plt.gca().add_patch(seeing)
-            else:
-                plot_random_lens()
+    def plot_lens_random_date(self, lensID = 7176527, convolve=False):
+        plotting.plot_lens_random_date(self, lensID, convolve)
 
-    def seeingToSig(seeing):
+"""
+    def determineAlpha(self, mag_ratio):
+        if(mag_ratio>1):
+            quasar_alpha = 1/mag_ratio
+            lens_alpha = 1
+        else:
+            quasar_alpha = 1
+            lens_alpha = mag_ratio
+        return quasar_alpha, lens_alpha
+
+    def determineScale(self, lensX, sourceX, lensY, sourceY):
+        if (abs(lensX[0]+sourceX))>(abs(sourceX)):
+            plotX = abs(lensX[0]+sourceX)
+        else:
+            plotX = abs(sourceX)
+        if (abs(lensY[0]+sourceY))>(abs(sourceY)):
+            plotY = abs(lensY[0]+sourceY)
+        else:
+            plotY = abs(sourceY)
+        return plotX, plotY
+
+    def plotFigureOnMatplotlib(self, currObs, convolve, quasar_alpha, lens_alpha, sourceX, sourceY, lensX, lensY, plotX, plotY):
+        fig = plt.figure(figsize=(4, 16))
+        scale_factor = 2
+        for i in range(4):
+            if(currObs[1]=='g'):
+                circleColor = 'g'
+            else:
+                circleColor = 'r'
+            plotNumStr = "4"+"1"+str(i+1)
+            plotNum = int(plotNumStr)
+            fig.subplots_adjust(top=1.5)
+            sub = fig.add_subplot(plotNum)
+            #sub.set_ylim(-plotY-2*currObs[2]/scale_factor, plotY+2*currObs[2]/scale_factor)
+            sub.set_ylim(-3, 3)
+            #sub.set_xlim(-plotX-2*currObs[2]/scale_factor, plotX+2*currObs[2]/scale_factor)
+            sub.set_xlim(-3, 3)
+            sub.set_xlabel('xPosition')
+            sub.set_ylabel('yPosition')
+            source = plt.Circle((sourceX, sourceY), radius=currObs[2]/scale_factor, alpha=quasar_alpha, fc=circleColor, linewidth=0)
+            if(convolve):
+                # For now, we just let lensFWHM to be zero.
+                #lensFWHM = currLens['REFF']
+                lensFWHM = 0.0
+                lens = plt.Circle((lensX[i]+sourceX, lensY[i]+sourceY), radius=convolve(currObs[2], lensFWHM)/scale_factor, alpha=lens_alpha, fc=circleColor, linewidth=0)
+            else:
+                lens = plt.Circle((lensX[i]+sourceX, lensY[i]+sourceY), radius=currObs[2]/scale_factor, alpha=lens_alpha, fc=circleColor, linewidth=0)
+            fig.gca().add_patch(lens)
+            fig.gca().add_patch(source)
+            sub.set_title('Observation ' + str(i+1) + ' with ' + 'filter ' + currObs[1] + ' on MJD ' + str(currObs[0]))
+            #seeing = plt.Circle(((-plotY-2*currObs[2])/scale_factor, (plotX-2*currObs[2])/scale_factor), radius=currObs[2]/scale_factor, alpha=0.1, fc='k')
+            seeing = plt.Circle((-2.5, -2.5), radius=currObs[2]/scale_factor, alpha=0.1, fc='k')
+            sub.legend((source, seeing, lens),('source', 'seeing', 'lens'))
+            plt.gca().add_patch(seeing)
+
+
+    def drawPlot(self, currObs, lensID = 7176527, convolve = False):
+        currLens = self.lens.get_lens(lensID)
+        #obsHist has MJD Filter FWHM 5sigmag
+        filterQuasar = currObs[1] + '_SDSS_quasar'
+        filterLens = currObs[1] + '_SDSS_lens'
+        lens_mag = currLens[filterLens]
+        quasar_mag = currLens[filterQuasar]
+        mag_ratio = math.pow(2.5, -lens_mag+quasar_mag)
+        quasar_alpha, lens_alpha = self.determineAlpha(mag_ratio)
+        sourceX = currLens['XSRC'][0]
+        sourceY = currLens['YSRC'][0]
+        lensX = currLens['XIMG'][0]
+        lensY = currLens['YIMG'][0]
+        plotX, plotY = self.determineScale(lensX, sourceX, lensY, sourceY)
+        self.plotFigureOnMatplotlib(currObs, convolve, quasar_alpha, lens_alpha, sourceX, sourceY, lensX, lensY, plotX, plotY)
+
+    def convolve(obsFWHM, initialFWHM = 0.0):
+        seeing = obsFWHM
+        seeingsigma = seeingToSig(seeing)
+        initSigma = seeingToSig(initialFWHM)
+        convolveSigma = seeingsigma + initSigma
+        return convolveSigma
+    
+    def seeingToSig(self, seeing):
         return seeing / np.sqrt(8 * np.log(2))
 
-    def sigmaTofwhm(sigma):
+    def sigmaTofwhm(self, sigma):
         return sigma * np.sqrt(8 * np.log(2))
+"""
