+++ python/desc/slrealizer/__init__.py
@@ -1,1 +1,2 @@
 from SLRealizer import *
+from plotting import *
