--- python/test/setup.py
@@ -1,19 +1,0 @@
-from setuptools import setup
-
-setup(
-      name="SLRealizer",
-      version="0.1",
-      author="Phil Marshall",
-      author_email="drphilmarshall@gmail.com",
-      url="https://github.com/LSSTDESC/SLRealizer",
-      packages=['slrealizer'],
-      description='Strong Lensed Quasar Realizer',
-      classifiers=[
-                   "Development Status :: 4 - Beta",
-                   "License :: OSI Approved :: MIT License",
-                   "Intended Audience :: Developers",
-                   "Intended Audience :: Science/Research",
-                   "Operating System :: OS Independent",
-                   "Programming Language :: Python",
-                   ],
-      )
