--- python/test/slrealizer/SConscript
@@ -1,4 +1,0 @@
-# -*- python -*-
-from lsst.sconsUtils import scripts
-# Uncomment the following line to enable swig builds.
-#scripts.BasicSConscript.python(['slrealizer'])
