--- python/test/slrealizer/__init__.py
@@ -1,1 +1,0 @@
-from SLRealizer import *
